% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_skeleton.R
\name{mui_skeleton}
\alias{mui_skeleton}
\title{Rating Input}
\usage{
mui_skeleton(
  inputId,
  default = NULL,
  variant = "rectangular",
  animation = "pulse",
  ...
)
}
\arguments{
\item{inputId}{Input id}

\item{default}{Default value}

\item{variant}{shape of skeleton c("text","rectangular", "circular").  default="rectangular"}

\item{animation}{The animation. If false the animation effect is disabled. c("pulse", "wave"). default="pulse"}

\item{...}{other props/arguments}
}
\description{
UI Shiny Input for Rating
}
\examples{

mui_skeleton("sk_1",variant = "rectangular", height = 400, width = "100\%")


}
\seealso{
\href{https://mui.com/api/skeleton/}{Additional arguments/props}
}
