% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_timeline.R
\name{mui_timelineInput}
\alias{mui_timelineInput}
\title{Rating Input}
\usage{
mui_timelineInput(
  inputId,
  value = c("First", "Second", "Third"),
  position = "right"
)
}
\arguments{
\item{inputId}{Input id}

\item{value}{list of items}

\item{position}{text position c("right", "left", "alternate")}
}
\description{
UI Shiny Input for Rating
}
\examples{

mui_timelineInput("timeline_1", position = "alternate")



}
