% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_transferlist.R
\name{mui_transferList}
\alias{mui_transferList}
\title{Transfer List UI}
\usage{
mui_transferList(
  inputId,
  default = NULL,
  left_list = c("A", "B", "C"),
  right_list = c("D", "E", "F"),
  left_title = "Choices",
  right_title = "Chosen",
  direction = "row",
  justifyContent = "center"
)
}
\arguments{
\item{inputId}{Input id}

\item{default}{Default value (will be the list on the right by default)}

\item{left_list}{Values of the list on the left (vector or list) (must be initialized with at least two values)}

\item{right_list}{Values of the list on the right (vector or list) (must be initialized with at least two values)}

\item{left_title}{The title of the list on the left (default  = "Choices")}

\item{right_title}{The title of the list on the right (default = "Chosen")}

\item{direction}{The layout of the lists c("row", "column"). Default = "row"}

\item{justifyContent}{justifyContent prop passed to the main grid. options: flex-start, center, flex-end, space-between, space-around, space-evenly}
}
\description{
UI Shiny Transfer List
}
\examples{

mui_transferList("transferList_1", left_list = names(mtcars),
right_list = names(iris))


}
\seealso{
\href{https://mui.com/api/transfer-list/}{Additional arguments/props}
}
