% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_checkbox.R
\name{updateMui_checkboxInput}
\alias{updateMui_checkboxInput}
\title{MUI Rating Updated Input}
\usage{
updateMui_checkboxInput(session, inputId, value, configuration = NULL)
}
\description{
Updating UI Shiny Input for Checkbox
}
