% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_timeline.R
\name{updateMui_timelineInput}
\alias{updateMui_timelineInput}
\title{Rating Update Input}
\usage{
updateMui_timelineInput(session, inputId, value, configuration = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{Input id}

\item{value}{Input value}

\item{configuration}{Other arguments/props}
}
\description{
Updating UI Shiny Input for Rating
}
