% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_transferlist.R
\name{updateMui_transferList}
\alias{updateMui_transferList}
\title{Transfer List}
\usage{
updateMui_transferList(session, inputId, value, configuration = NULL)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{Input id}

\item{value}{Input value}

\item{configuration}{Other arguments/props}
}
\description{
Updating Transfer List
}
