% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_timeline.R
\name{mui_timelineInput}
\alias{mui_timelineInput}
\title{Timeline Input}
\usage{
mui_timelineInput(
  inputId,
  value = NULL,
  position = "right",
  subvalues = NULL,
  ...
)
}
\arguments{
\item{inputId}{Input id}

\item{value}{list of items}

\item{position}{text position c("right", "left", "alternate")}

\item{subvalues}{subtext of each list title}

\item{...}{other options}
}
\description{
UI Shiny Input for Timeline
}
\examples{

mui_timelineInput("timeline_1", value = c("List 1", "List 2", "List 3"), position = "alternate")



}
