% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mui_rating.R
\name{updateMui_ratingInput}
\alias{updateMui_ratingInput}
\title{MUI Rating Updated Input}
\usage{
updateMui_ratingInput(session, inputId, value, configuration = NULL)
}
\description{
Updating UI Shiny Input for Rating
}
